backend/requirements.txt
flask
flask-cors
python-dotenv
requests

# Optional: If you integrate uagents from Fetch.ai later
# uagents
